project images faker seed

+++ database/migrations/2021_08_24_134917_create_project_images_table.php
@@ -15,7 +15,7 @@ class CreateProjectImagesTable extends Migration
     {
         Schema::create('project_images', function (Blueprint $table) {
             $table->id();
-            $table->integer('project_id')->unsigned();
+            $table->bigInteger('project_id')->unsigned();
             $table->char('image');
             $table->timestamps();
